requirements.txt
alembic~=1.11.2
asyncpg
fastapi[all]~=0.101.1
fastapi-users[sqlalchemy]
python-dotenv~=1.0.0
sqlalchemy~=2.0.19
pydantic~=2.1.1
pyjwt~=2.8.0
cryptography~=39.0.1
pandas~=1.5.3
websockets~=11.0.3
requests~=2.29.0
pymexc~=1.0.9
aiohttp~=3.8.1
numpy~=1.23.5
peewee~=3.15.3
pip~=23.2.1
toml~=0.10.2
zlib~=1.2.13
wheel~=0.38.4
brotli~=1.0.9
openssl~=1.1.1v
py~=1.11.0
lxml~=4.9.2
tornado~=6.2
keyring~=23.13.1
setuptools~=66.0.0
cython~=0.29.24
future~=0.18.3
docutils~=0.18.1
pyqt5~=5.15.9
ntplib~=0.4.0
pywin32~=305
scikit-learn~=1.2.2
mplfinance~=0.12.9b7
psutil~=5.9.0
wmi~=1.5.1
pyqtgraph~=0.13.1
pybit~=2.4.1